Overdispersed starting values diagnostic for multiple imputation

A visual diagnostic of EM convergence from multiple overdispersed starting values for an output from amelia.

disperse(
  output,
  m = 5,
  dims = 1,
  p2s = 0,
  frontend = FALSE,
  ...,
  xlim = NULL,
  ylim = NULL
)

Arguments

output

output from the function amelia.

m

the number of EM chains to run from overdispersed starting values.

dims

the number of principle components of the parameters to display and assess convergence on (up to 2).

p2s

an integer that controls printing to screen. 0 (default) indicates no printing, 1 indicates normal screen output and 2 indicates diagnostic output.

frontend

a logical value used internally for the Amelia GUI.

...

further graphical parameters for the plot.

xlim

limits of the plot in the horizontal dimension.

ylim

limits of the plot in vertical dimension.

Details

This function tracks the convergence of m EM chains which start from various overdispersed starting values. This plot should give some indication of the sensitivity of the EM algorithm to the choice of starting values in the imputation model in output. If all of the lines converge to the same point, then we can be confident that starting values are not affecting the EM algorithm.

As the parameter space of the imputation model is of a high-dimension, this plot tracks how the first (and second if dims is 2) principle component(s) change over the iterations of the EM algorithm. Thus, the plot is a lower dimensional summary of the convergence and is subject to all the drawbacks inherent in said summaries.

For dims==1, the function plots a horizontal line at the position where the first EM chain converges. Thus, we are checking that the other chains converge close to that horizontal line. For dims==2, the function draws a convex hull around the point of convergence for the first EM chain. The hull is scaled to be within the tolerance of the EM algorithm. Thus, we should check that the other chains end up in this hull.

See also