Transform imputed datasets from Amelia objects

Updates the imputed datasets from an amelia output with the specified transformations.

# S3 method for amelia
transform(`_data`, ...)

Arguments

_data

an object of class "amelia"; typically output from the function amelia.

...

further arguments of the form tag = value.

Value

An object of class amelia with its imputations and missMatrix values updated according to the transformations. In addition, each of the calls to transform.amelia are stored in

Details

The ... arugments to transform.amelia are expressions of the form tag = value, where tag is the variable that is being updated or created and value is an expression that is a function of the variables in the imputed datasets. For instance, if you wanted to create an interaction of two imputed variables, you could have one argument be intervar = var1 * var2. This would either update the current variable intervar in the imputed data or append a new variable called intervar to the imputed datasets.

See also