Execute commands within each imputed data set

Evaluate an R expression in the environments constructed from the imputed data sets of a call to amelia function.

# S3 method for amelia
with(data, expr, ...)

Arguments

data

imputation output from the amelia funtion.

expr

expression to evaluate in each imputed data set in data.

...

arguments to be passed to (future) methods.

Value

a list the same length as data$imputations that contains the output of the expression as evaluated in each imputed data set of data.

Author

Matt Blackwell

Examples

data(africa)
a.out <- amelia(x = africa, cs = "country", ts = "year", logs =
"gdp_pc")
#> -- Imputation 1 --
#> 
#>   1  2  3
#> 
#> -- Imputation 2 --
#> 
#>   1  2  3
#> 
#> -- Imputation 3 --
#> 
#>   1  2  3
#> 
#> -- Imputation 4 --
#> 
#>   1  2
#> 
#> -- Imputation 5 --
#> 
#>   1  2  3
#> 

imp.mods <- with(a.out, lm(gdp_pc ~ infl + trade))

mi.combine(imp.mods, conf.int = TRUE)
#> # A tibble: 3 × 10
#>   term   estim…¹ std.e…² stati…³  p.value     df       r miss.…⁴ conf.…⁵ conf.…⁶
#>   <chr>    <dbl>   <dbl>   <dbl>    <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 (Inte…    1.78   83.4   0.0213 9.83e- 1 1.34e4 0.0176  0.0174   165.    -162. 
#> 2 infl     -6.88    1.67 -4.13   2.00e+ 0 1.03e6 0.00197 0.00197   -3.61   -10.1
#> 3 trade    18.4     1.24 14.9    1.04e-49 8.18e3 0.0226  0.0223    20.9     16.0
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​miss.info, ⁵​conf.low, ⁶​conf.high