Combine results from analyses on imputed data sets

Combine results from statistical models run on multiply imputed data sets using the so-called Rubin rules.

mi.combine(x, conf.int = FALSE, conf.level = 0.95)

Arguments

x

List of output from statistical models estimated on different imputed data sets, as outputted by with(a.out, expr) where a.out is the output of a call to amelia.

conf.int

Logical indicating if confidence intervals should be computed for each quantity of interest (default is FALSE).

conf.level

The confidence level to use for the confidence interval if conf.level = TRUE. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

Value

Returns a tibble that contains:

term

Name of the coefficient or parameter.

estimate

Estimate of the parameter, averagine across imputations.

std.error

Standard error of the estimate, accounting for imputation uncertainty.

statistic

Value of the t-statistic for the estimated parameter.

p.value

p-value associated with the test of a null hypothesis that the true coefficient is zero. Uses the t-distribution with an imputation-adjusted degrees of freedom.

df

Imputation-adjusted degrees of freedom for each parameter.

r

Relative increase in variance due to nonresponse.

miss.info

Estimated fraction of missing information.

conf.low

Lower bound of the estimated confidence interval. Only present if conf.int = TRUE.

conf.high

Upper bound of the estimated confidence interval. Only present if conf.int = TRUE.

Author

Matt Blackwell